os.gentoo.conf: uncomment the clamd_socket by default.
The Gentoo clamav package has LocalSocket enabled (and set to the
expected value) by default, so it makes sense to have clamd_socket
uncommented in the Gentoo configuration. The Gentoo package (of
clamav-unofficial-sigs) now uses os.gentoo.conf, so having this
uncommented means that the ebuild doesn't have to patch os.gentoo.conf
before installing it.

If users install clamav-unofficial-sigs from the ebuild, it will
ensure that the additional dependencies are pulled in. Otherwise,
their absense is not fatal, and the clamd_socket line can always be
commented-out for users who wish to avoid both the package manager and
the perl dependencies.